Hazelnut Theme.

+++ Hazel/src/Hazel/Debug/Instrumentor.h
@@ -11,7 +11,7 @@ namespace Hazel {
 	struct ProfileResult {
 		std::string Name;
 		long long Start, End;
-		uint32_t ThreadID;
+		uint64_t ThreadID;
 	};
 
 	struct InstrumentationSession {
@@ -95,7 +95,7 @@ namespace Hazel {
 			long long start = std::chrono::time_point_cast<std::chrono::microseconds>(m_StartTimepoint).time_since_epoch().count();
 			long long end = std::chrono::time_point_cast<std::chrono::microseconds>(endTimepoint).time_since_epoch().count();
 
-			uint32_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
+			uint64_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
 			Instrumentor::Get().WriteProfile({ m_Name, start, end, threadID });
 
 			m_Stopped = true;
